Update meshKronPoly.m
According to the documentation for meshgrid():
"[X,Y] = meshgrid(x,y) returns 2-D grid coordinates based on the coordinates contained in vectors x and y. X is a matrix where each row is a copy of x, and Y is a matrix where each column is a copy of y. The grid represented by the coordinates X and Y has length(y) rows and length(x) columns."

In lines 30 and 31, the indices needed to be flipped so i=1:n2 and j=1:n1. This is because X and Y are both n2 by n1 matrices due to meshgrid().