Correct auto-transposing in kronPolyEval.m
A recent commit added the functionality to automatically transpose the coefficients c if they were not already transposed; however, this introduced a bug where the coefficients were transposed even if they weren't supposed to be. This was based on the assumption that the coefficients would have dimension 1 in one dimension; however, this is not always the case, as in when we are using kronPolyEval to perform a change of coordinates for example. This commit basically does a more robust check to see if the coefficients should be transposed; instead of transposing based on one of the dimensions being 1, we check to see if the dimensions are appropriate to perform the desired multiplications. For the sake of speed, I only check the last coefficient, but we could do something more robust if necessary.